Offset hard-cap file growth with extractions; track CmuxEventBus budget
The simulator wiring adds a few lines to four files that sit over the
file-length hard cap, where CI rejects any net growth. Offset each with a
small extraction into a new file (the sanctioned pattern for these files):
- ContentView.swift -> ContentView+SurfaceKindMetadata.swift (the palette
  surface-kind label/keyword switches)
- Workspace.swift -> Workspace+CloseConfirmation.swift
- TerminalController.swift -> TerminalController+V2CallResult.swift
- CLI/cmux.swift -> CMUXCLI+V2Output.swift (formatHandle / printV2Payload)

Also adds the missing budget row for Sources/CmuxEventBus.swift (619 lines),
which crossed the tracking threshold on main without a TSV entry and fails
the budget gate for any branch that diffs against it.
